requirement.txt
flask
pandas
matplotlib
python-dotenv
google-genai
